feat: add NestJS boilerplate
