+++ src/common/dto/answer.dto.ts
@@ -1,0 +1,6 @@
+export interface AnswerDto {
+  '6_month_average_income': number;
+  min_balance: number;
+  max_balance: number;
+  '3_years_activity': boolean;
+}
