+++ src/common/ro/transaction.ro.ts
@@ -1,0 +1,8 @@
+export interface TransactionRo {
+  timestamp: string;
+  transaction_type: string;
+  transaction_category: string;
+  amount: number;
+  currency: string;
+  status: string;
+}
