--- src/ro/transaction.ro.ts
@@ -1,8 +1,0 @@
-export interface TransactionRo {
-    timestamp: string;
-    transaction_type: string;
-    transaction_category: string;
-    amount: number;
-    currency: string;
-    status: string;
-}
